refactor: move to unplugin instead
