+++ examples/rollup-express-web-atomic/package.json
@@ -1,11 +1,9 @@
 {
   "scripts": {
     "prebuild": "rimraf dist",
-    "dev": "pnpm run build:server; pnpm run build:client; concurrently 'pnpm:build:* -w' 'pnpm:dev:nodemon'",
-    "build:server": "rollup -c --config-server",
-    "build:client": "rollup -c --config-client",
+    "dev": "pnpm run build; concurrently 'pnpm:build:* -w' 'pnpm:dev:nodemon'",
+    "build": "rollup -c",
     "dev:nodemon": "nodemon --watch src --watch dist --ext js dist/server.js",
-    "build": "pnpm run build:server; pnpm run build:client;",
     "fix": "prettier --write ."
   },
   "devDependencies": {
