+++ package.json
@@ -50,20 +50,21 @@
     "@babel/core": "^7.21.3",
     "@babel/generator": "^7.21.1",
     "@babel/parser": "^7.21.1",
-    "@babel/preset-typescript": "^7.21.0"
+    "@babel/preset-typescript": "^7.21.0",
+    "esbuild": "^0.17.16"
   },
   "devDependencies": {
     "@barelyhuman/prettier-config": "^1.1.0",
     "@types/babel__core": "^7.20.0",
     "@types/node": "^18.14.0",
     "bumpp": "^9.0.0",
-    "esbuild": "^0.17.16",
     "fast-glob": "^3.2.12",
     "husky": "^8.0.3",
     "lint-staged": "^13.1.2",
     "prettier": "^2.8.4",
     "tsup": "^6.6.3",
     "typescript": "^4.9.5",
+    "unplugin": "^1.3.1",
     "vite": "^4.1.0",
     "vitest": "^0.32.0"
   },
@@ -72,10 +73,10 @@
     "vite": "^4.1.0"
   },
   "peerDependenciesMeta": {
-    "esbuild": {
+    "vite": {
       "optional": true
     },
-    "vite": {
+    "rollup": {
       "optional": true
     }
   },
