+++ tsup.config.ts
@@ -4,5 +4,7 @@ export default <Options>{
   entryPoints: ['src/*.ts'],
   clean: true,
   format: ['cjs', 'esm'],
+  platform: 'node',
+  external: ['path'],
   dts: true,
 }
